HiSTBAndroidV600R001C00SPC062

+++ bigfish/sdk/source/msp/api/jpeg/src/src_6b/jccolor.c
@@ -439,7 +439,7 @@ grayscale_convert (j_compress_ptr cinfo,
 
 
 #if defined(CONFIG_JPEG_ADD_GOOGLEFUNCTION) && defined(ENABLE_ANDROID_NULL_CONVERT)
-	typedef unsigned long UINT32;
+    typedef unsigned int UINT32; /** support 64bit **/
 	#define B0(n)   ((n) & 0xFF)
 	#define B1(n)   (((n) >> 8) & 0xFF)
 	#define B2(n)   (((n) >> 16) & 0xFF)
